added removed dali base reader
