+++ dlio_benchmark/reader/dali_npy_reader.py
@@ -20,7 +20,7 @@ from time import time
 
 import nvidia.dali.fn as fn
 from dlio_benchmark.common.constants import MODULE_DATA_READER
-from dlio_benchmark.reader.dali_base_reader import DaliBaseReader
+from dlio_benchmark.dlio_benchmark.reader.reader_handler import FormatReader
 from dlio_benchmark.utils.utility import utcnow
 from dlio_benchmark.common.enumerations import DatasetType, Shuffle
 import nvidia.dali.tfrecord as tfrec
@@ -29,13 +29,13 @@ from dlio_profiler.logger import dlio_logger as PerfTrace, fn_interceptor as Pro
 dlp = Profile(MODULE_DATA_READER)
 
 
-class DaliNPYReader(DaliBaseReader):
+class DaliNPYReader(NativeDaliBaseReader):
     @dlp.log_init
-    def __init__(self, dataset_type):
-        super().__init__(dataset_type)
+    def __init__(dataset_type, thread_index, epoch):
+        super().__init__(dataset_type, thread_index)
 
     @dlp.log
-    def _load(self):
+    def read(self):
         logging.debug(
             f"{utcnow()} Reading {len(self.file_list)} files rank {self._args.my_rank}")
         random_shuffle = False
@@ -61,8 +61,22 @@ class DaliNPYReader(DaliBaseReader):
                                    prefetch_queue_depth=prefetch_size, initial_fill=initial_fill,
                                    random_shuffle=random_shuffle, seed=seed, shuffle_after_epoch=seed_change_epoch,
                                    stick_to_shard=stick_to_shard, pad_last_batch=True)
+
+        dataset = self._preprocess(dataset)
+        dataset = self._resize(dataset)
+        return dataset
+    
+
+    @dlp.log
+    def _preprocess(self, dataset):
+        if self._args.preprocess_time != 0. or self._args.preprocess_time_stdev != 0.:
+            t = np.random.normal(self._args.preprocess_time, self._args.preprocess_time_stdev)
+            sleep(max(t, 0.0))
         return dataset
 
+    @dlp.log
+    def _resize(self, dataset):
+        return nvidia.dali.fn.reshape(dataset, shape=[self._args.max_dimension, self._args.max_dimension])
     @dlp.log
     def finalize(self):
         pass
