+++ dlio_benchmark/reader/reader_factory.py
@@ -46,14 +46,14 @@ class ReaderFactory(object):
         elif type == FormatType.JPEG or type == FormatType.PNG:
             if _args.data_loader == DataLoaderType.NATIVE_DALI:
                 from dlio_benchmark.reader.dali_image_reader import DaliImageReader
-                return DaliImageReader(dataset_type)
+                return DaliImageReader(dataset_type, thread_index, epoch_number)
             else:
                 from dlio_benchmark.reader.image_reader import ImageReader
                 return ImageReader(dataset_type, thread_index, epoch_number)   
         elif type == FormatType.NPY:
             if _args.data_loader == DataLoaderType.NATIVE_DALI:
                 from dlio_benchmark.reader.dali_npy_reader import DaliNPYReader
-                return DaliNPYReader(dataset_type)
+                return DaliNPYReader(dataset_type, thread_index, epoch_number)
             else:
                 from dlio_benchmark.reader.npy_reader import NPYReader
                 return NPYReader(dataset_type, thread_index, epoch_number)                         
@@ -67,7 +67,7 @@ class ReaderFactory(object):
         elif type == FormatType.TFRECORD:
             if _args.data_loader == DataLoaderType.NATIVE_DALI: 
                 from dlio_benchmark.reader.dali_tfrecord_reader import DaliTFRecordReader
-                return DaliTFRecordReader(dataset_type)
+                return DaliTFRecordReader(dataset_type, thread_index, epoch_number)
             else:
                 from dlio_benchmark.reader.tf_reader import TFReader
                 return TFReader(dataset_type, thread_index, epoch_number)
